Conflicts:
	efunctions/ebegin
	efunctions/eend
	efunctions/eerror
	efunctions/eerrorn
	efunctions/einfo
	efunctions/einfon
	efunctions/eval_ecolors
	efunctions/ewarn
	efunctions/ewarnn
	efunctions/ewend
	test.sh

+++ install.sh
@@ -1,5 +1,13 @@
 HERE=$(readlink -f "$0")
 DIR=$(dirname "$HERE")
 SCRIPT="$DIR/functions.sh"
-echo "Installing $SCRIPT into /etc/init.d"
+
+source "$SCRIPT"
+
+if [ -f /etc/init.d/functions.sh ]; then
+  "$DIR/efunctions/eerror" "Cannot install efunctions: file already exists"
+  exit
+fi
+
+"$DIR/efunctions/einfo" "Installing $SCRIPT into /etc/init.d"
 sudo ln -s "$SCRIPT" /etc/init.d/functions.sh
